Moved character equipment, job data and skills to their own DAO classes
