+++ Arrowgene.Ddon.Database/Sql/Core/DdonSqlDbCharacter.cs
@@ -15,7 +15,7 @@ namespace Arrowgene.Ddon.Database.Sql.Core
             "version", "account_id", "first_name", "last_name", "created", "job", "jewelry_slot_num", "my_pawn_slot_num", "rental_pawn_slot_num", "hide_equip_head", "hide_equip_lantern", "hide_equip_head_pawn", "hide_equip_lantern_pawn", "arisen_profile_share_range"
         };
 
-                private static readonly string[] CDataEditInfoFields = new string[]
+        private static readonly string[] CDataEditInfoFields = new string[]
         {
             "character_id", "sex", "voice", "voice_pitch", "personality", "speech_freq", "body_type", "hair", "beard", "makeup", "scar",
             "eye_preset_no", "nose_preset_no", "mouth_preset_no", "eyebrow_tex_no", "color_skin", "color_hair",
@@ -90,73 +90,6 @@ namespace Arrowgene.Ddon.Database.Sql.Core
         private static readonly string SqlUpdateCharacterArisenProfile = $"UPDATE `ddon_character_arisen_profile` SET {BuildQueryUpdate(CDataArisenProfileFields)} WHERE `character_id` = @character_id;";
         private static readonly string SqlSelectCharacterArisenProfile = $"SELECT {BuildQueryField(CDataArisenProfileFields)} FROM `ddon_character_arisen_profile` WHERE `character_id` = @character_id;";
         private const string SqlDeleteCharacterArisenProfile = "DELETE FROM `ddon_character_arisen_profile` WHERE `character_id`=@character_id;";
-
-
-        private static readonly string[] CDataCharacterJobDataFields = new string[]
-        {
-            "character_id", "job", "exp", "job_point", "lv", "atk", "def", "m_atk", "m_def", "strength", "down_power", "shake_power", 
-            "stun_power", "consitution", "guts", "fire_resist", "ice_resist", "thunder_resist", "holy_resist", "dark_resist", "spread_resist", 
-            "freeze_resist", "shock_resist", "absorb_resist", "dark_elm_resist", "poison_resist", "slow_resist", "sleep_resist", "stun_resist", 
-            "wet_resist", "oil_resist", "seal_resist", "curse_resist", "soft_resist", "stone_resist", "gold_resist", "fire_reduce_resist", 
-            "ice_reduce_resist", "thunder_reduce_resist", "holy_reduce_resist", "dark_reduce_resist", "atk_down_resist", "def_down_resist", 
-            "m_atk_down_resist", "m_def_down_resist"
-        };
-
-        private readonly string SqlInsertCharacterJobData = $"INSERT INTO `ddon_character_job_data` ({BuildQueryField(CDataCharacterJobDataFields)}) VALUES ({BuildQueryInsert(CDataCharacterJobDataFields)});";
-        private readonly string SqlReplaceCharacterJobData = $"INSERT OR REPLACE INTO `ddon_character_job_data` ({BuildQueryField(CDataCharacterJobDataFields)}) VALUES ({BuildQueryInsert(CDataCharacterJobDataFields)});";
-        private static readonly string SqlUpdateCharacterJobData = $"UPDATE `ddon_character_job_data` SET {BuildQueryUpdate(CDataCharacterJobDataFields)} WHERE `character_id` = @character_id AND `job` = @job;";
-        private static readonly string SqlSelectCharacterJobData = $"SELECT {BuildQueryField(CDataCharacterJobDataFields)} FROM `ddon_character_job_data` WHERE `character_id` = @character_id AND `job` = @job;";
-        private static readonly string SqlSelectCharacterJobDataByCharacter = $"SELECT {BuildQueryField(CDataCharacterJobDataFields)} FROM `ddon_character_job_data` WHERE `character_id` = @character_id;";
-        private const string SqlDeleteCharacterJobData = "DELETE FROM `ddon_character_job_data` WHERE `character_id`=@character_id AND `job` = @job;";
-        
-
-        private static readonly string[] CDataEquipItemInfoFields = new string[]
-        {
-            "character_id", "job", "item_id", "equip_type", "equip_slot", "color", "plus_value"
-        };
-
-        private readonly string SqlInsertEquipItemInfo = $"INSERT INTO `ddon_equip_item_info` ({BuildQueryField(CDataEquipItemInfoFields)}) VALUES ({BuildQueryInsert(CDataEquipItemInfoFields)});";
-        private readonly string SqlReplaceEquipItemInfo = $"INSERT OR REPLACE INTO `ddon_equip_item_info` ({BuildQueryField(CDataEquipItemInfoFields)}) VALUES ({BuildQueryInsert(CDataEquipItemInfoFields)});";
-        private static readonly string SqlUpdateEquipItemInfo = $"UPDATE `ddon_equip_item_info` SET {BuildQueryUpdate(CDataEquipItemInfoFields)} WHERE `character_id` = @character_id AND `job` = @job AND `equip_type`=@equip_type AND `equip_slot`=@equip_slot;";
-        private static readonly string SqlSelectEquipItemInfo = $"SELECT {BuildQueryField(CDataEquipItemInfoFields)} FROM `ddon_equip_item_info` WHERE `character_id` = @character_id AND `job` = @job AND `equip_type`=1;";
-        private static readonly string SqlSelectVisualEquipItemInfo = $"SELECT {BuildQueryField(CDataEquipItemInfoFields)} FROM `ddon_equip_item_info` WHERE `character_id` = @character_id AND `job` = @job AND `equip_type`=2;";
-        private const string SqlDeleteEquipItemInfo = "DELETE FROM `ddon_equip_item_info` WHERE `character_id`=@character_id AND `job`=@job AND `equip_type`=@equip_type AND `equip_slot`=@equip_slot;";
-
-
-        private static readonly string[] CDataEquipJobItemFields = new string[]
-        {
-            "character_id", "job", "job_item_id", "equip_slot_no"
-        };
-
-        private readonly string SqlInsertEquipJobItem = $"INSERT INTO `ddon_equip_job_item` ({BuildQueryField(CDataEquipJobItemFields)}) VALUES ({BuildQueryInsert(CDataEquipJobItemFields)});";
-        private readonly string SqlReplaceEquipJobItem = $"INSERT OR REPLACE INTO `ddon_equip_job_item` ({BuildQueryField(CDataEquipJobItemFields)}) VALUES ({BuildQueryInsert(CDataEquipJobItemFields)});";
-        private static readonly string SqlUpdateEquipJobItem = $"UPDATE `ddon_equip_job_item` SET {BuildQueryUpdate(CDataEquipJobItemFields)} WHERE `character_id` = @character_id AND `job` = @job AND `equip_slot_no`=@equip_slot_no;";
-        private static readonly string SqlSelectEquipJobItem = $"SELECT {BuildQueryField(CDataEquipJobItemFields)} FROM `ddon_equip_job_item` WHERE `character_id` = @character_id AND `job` = @job;";
-        private const string SqlDeleteEquipJobItem = "DELETE FROM `ddon_equip_job_item` WHERE `character_id`=@character_id AND `job`=@job AND `equip_slot_no`=@equip_slot_no;";
-
-
-        private static readonly string[] CDataNormalSkillParamFields = new string[]
-        {
-            "character_id", "job", "skill_no", "index", "pre_skill_no"
-        };
-
-        private readonly string SqlInsertNormalSkillParam = $"INSERT INTO `ddon_normal_skill_param` ({BuildQueryField(CDataNormalSkillParamFields)}) VALUES ({BuildQueryInsert(CDataNormalSkillParamFields)});";
-        private readonly string SqlReplaceNormalSkillParam = $"INSERT OR REPLACE INTO `ddon_normal_skill_param` ({BuildQueryField(CDataNormalSkillParamFields)}) VALUES ({BuildQueryInsert(CDataNormalSkillParamFields)});";
-        private static readonly string SqlUpdateNormalSkillParam = $"UPDATE `ddon_normal_skill_param` SET {BuildQueryUpdate(CDataNormalSkillParamFields)} WHERE `character_id` = @character_id AND `job` = @job AND `skill_no`=@skill_no;";
-        private static readonly string SqlSelectNormalSkillParam = $"SELECT {BuildQueryField(CDataNormalSkillParamFields)} FROM `ddon_normal_skill_param` WHERE `character_id` = @character_id;";
-        private const string SqlDeleteNormalSkillParam = "DELETE FROM `ddon_normal_skill_param` WHERE `character_id`=@character_id AND `job`=@job AND `skill_no`=@skill_no;";
-
-        private static readonly string[] CDataSetAcquirementParamFields = new string[]
-        {
-            "character_id", "job", "type", "slot_no", "acquirement_no", "acquirement_lv"
-        };
-
-        private readonly string SqlInsertSetAcquirementParam = $"INSERT INTO `ddon_set_acquirement_param` ({BuildQueryField(CDataSetAcquirementParamFields)}) VALUES ({BuildQueryInsert(CDataSetAcquirementParamFields)});";
-        private readonly string SqlReplaceSetAcquirementParam = $"INSERT OR REPLACE INTO `ddon_set_acquirement_param` ({BuildQueryField(CDataSetAcquirementParamFields)}) VALUES ({BuildQueryInsert(CDataSetAcquirementParamFields)});";
-        private static readonly string SqlUpdateSetAcquirementParam = $"UPDATE `ddon_set_acquirement_param` SET {BuildQueryUpdate(CDataSetAcquirementParamFields)} WHERE `character_id` = @character_id AND `job` = @job AND `slot_no`=@slot_no;";
-        private static readonly string SqlSelectCustomSkillsSetAcquirementParam = $"SELECT {BuildQueryField(CDataSetAcquirementParamFields)} FROM `ddon_set_acquirement_param` WHERE `character_id` = @character_id AND `job` <> 0";
-        private static readonly string SqlSelectAbilitiesSetAcquirementParam = $"SELECT {BuildQueryField(CDataSetAcquirementParamFields)} FROM `ddon_set_acquirement_param` WHERE `character_id` = @character_id AND `job` = 0;";
-        private const string SqlDeleteSetAcquirementParam = "DELETE FROM `ddon_set_acquirement_param` WHERE `character_id`=@character_id AND `job`=@job AND `slot_no`=@slot_no;";
 
 
         public bool CreateCharacter(Character character)
@@ -679,182 +612,6 @@ namespace Arrowgene.Ddon.Database.Sql.Core
             AddParameter(command, "@title_index", character.CharacterInfo.ArisenProfile.Title.Index);
             AddParameter(command, "@motion_id", character.CharacterInfo.ArisenProfile.MotionId);
             AddParameter(command, "@motion_frame_no", character.CharacterInfo.ArisenProfile.MotionFrameNo);
-        }
-        private CDataCharacterJobData ReadCharacterJobData(DbDataReader reader)
-        {
-            CDataCharacterJobData characterJobData = new CDataCharacterJobData();
-            characterJobData.Job = (JobId) GetByte(reader, "job");
-            characterJobData.Exp = GetUInt32(reader, "exp");
-            characterJobData.JobPoint = GetUInt32(reader, "job_point");
-            characterJobData.Lv = GetUInt32(reader, "lv");
-            characterJobData.Atk = GetUInt16(reader, "atk");
-            characterJobData.Def = GetUInt16(reader, "def");
-            characterJobData.MAtk = GetUInt16(reader, "m_atk");
-            characterJobData.MDef = GetUInt16(reader, "m_def");
-            characterJobData.Strength = GetUInt16(reader, "strength");
-            characterJobData.DownPower = GetUInt16(reader, "down_power");
-            characterJobData.ShakePower = GetUInt16(reader, "shake_power");
-            characterJobData.StunPower = GetUInt16(reader, "stun_power");
-            characterJobData.Consitution = GetUInt16(reader, "consitution");
-            characterJobData.Guts = GetUInt16(reader, "guts");
-            characterJobData.FireResist = GetByte(reader, "fire_resist");
-            characterJobData.IceResist = GetByte(reader, "ice_resist");
-            characterJobData.ThunderResist = GetByte(reader, "thunder_resist");
-            characterJobData.HolyResist = GetByte(reader, "holy_resist");
-            characterJobData.DarkResist = GetByte(reader, "dark_resist");
-            characterJobData.SpreadResist = GetByte(reader, "spread_resist");
-            characterJobData.FreezeResist = GetByte(reader, "freeze_resist");
-            characterJobData.ShockResist = GetByte(reader, "shock_resist");
-            characterJobData.AbsorbResist = GetByte(reader, "absorb_resist");
-            characterJobData.DarkElmResist = GetByte(reader, "dark_elm_resist");
-            characterJobData.PoisonResist = GetByte(reader, "poison_resist");
-            characterJobData.SlowResist = GetByte(reader, "slow_resist");
-            characterJobData.SleepResist = GetByte(reader, "sleep_resist");
-            characterJobData.StunResist = GetByte(reader, "stun_resist");
-            characterJobData.WetResist = GetByte(reader, "wet_resist");
-            characterJobData.OilResist = GetByte(reader, "oil_resist");
-            characterJobData.SealResist = GetByte(reader, "seal_resist");
-            characterJobData.CurseResist = GetByte(reader, "curse_resist");
-            characterJobData.SoftResist = GetByte(reader, "soft_resist");
-            characterJobData.StoneResist = GetByte(reader, "stone_resist");
-            characterJobData.GoldResist = GetByte(reader, "gold_resist");
-            characterJobData.FireReduceResist = GetByte(reader, "fire_reduce_resist");
-            characterJobData.IceReduceResist = GetByte(reader, "ice_reduce_resist");
-            characterJobData.ThunderReduceResist = GetByte(reader, "thunder_reduce_resist");
-            characterJobData.HolyReduceResist = GetByte(reader, "holy_reduce_resist");
-            characterJobData.DarkReduceResist = GetByte(reader, "dark_reduce_resist");
-            characterJobData.AtkDownResist = GetByte(reader, "atk_down_resist");
-            characterJobData.DefDownResist = GetByte(reader, "def_down_resist");
-            characterJobData.MAtkDownResist = GetByte(reader, "m_atk_down_resist");
-            characterJobData.MDefDownResist = GetByte(reader, "m_def_down_resist");
-            return characterJobData;
-        }
-
-        private void AddParameter(TCom command, uint characterId, CDataCharacterJobData characterJobData)
-        {
-            AddParameter(command, "character_id", characterId);
-            AddParameter(command, "job", (byte) characterJobData.Job);
-            AddParameter(command, "exp", characterJobData.Exp);
-            AddParameter(command, "job_point", characterJobData.JobPoint);
-            AddParameter(command, "lv", characterJobData.Lv);
-            AddParameter(command, "atk", characterJobData.Atk);
-            AddParameter(command, "def", characterJobData.Def);
-            AddParameter(command, "m_atk", characterJobData.MAtk);
-            AddParameter(command, "m_def", characterJobData.MDef);
-            AddParameter(command, "strength", characterJobData.Strength);
-            AddParameter(command, "down_power", characterJobData.DownPower);
-            AddParameter(command, "shake_power", characterJobData.ShakePower);
-            AddParameter(command, "stun_power", characterJobData.StunPower);
-            AddParameter(command, "consitution", characterJobData.Consitution);
-            AddParameter(command, "guts", characterJobData.Guts);
-            AddParameter(command, "fire_resist", characterJobData.FireResist);
-            AddParameter(command, "ice_resist", characterJobData.IceResist);
-            AddParameter(command, "thunder_resist", characterJobData.ThunderResist);
-            AddParameter(command, "holy_resist", characterJobData.HolyResist);
-            AddParameter(command, "dark_resist", characterJobData.DarkResist);
-            AddParameter(command, "spread_resist", characterJobData.SpreadResist);
-            AddParameter(command, "freeze_resist", characterJobData.FreezeResist);
-            AddParameter(command, "shock_resist", characterJobData.ShockResist);
-            AddParameter(command, "absorb_resist", characterJobData.AbsorbResist);
-            AddParameter(command, "dark_elm_resist", characterJobData.DarkElmResist);
-            AddParameter(command, "poison_resist", characterJobData.PoisonResist);
-            AddParameter(command, "slow_resist", characterJobData.SlowResist);
-            AddParameter(command, "sleep_resist", characterJobData.SleepResist);
-            AddParameter(command, "stun_resist", characterJobData.StunResist);
-            AddParameter(command, "wet_resist", characterJobData.WetResist);
-            AddParameter(command, "oil_resist", characterJobData.OilResist);
-            AddParameter(command, "seal_resist", characterJobData.SealResist);
-            AddParameter(command, "curse_resist", characterJobData.CurseResist);
-            AddParameter(command, "soft_resist", characterJobData.SoftResist);
-            AddParameter(command, "stone_resist", characterJobData.StoneResist);
-            AddParameter(command, "gold_resist", characterJobData.GoldResist);
-            AddParameter(command, "fire_reduce_resist", characterJobData.FireReduceResist);
-            AddParameter(command, "ice_reduce_resist", characterJobData.IceReduceResist);
-            AddParameter(command, "thunder_reduce_resist", characterJobData.ThunderReduceResist);
-            AddParameter(command, "holy_reduce_resist", characterJobData.HolyReduceResist);
-            AddParameter(command, "dark_reduce_resist", characterJobData.DarkReduceResist);
-            AddParameter(command, "atk_down_resist", characterJobData.AtkDownResist);
-            AddParameter(command, "def_down_resist", characterJobData.DefDownResist);
-            AddParameter(command, "m_atk_down_resist", characterJobData.MAtkDownResist);
-            AddParameter(command, "m_def_down_resist", characterJobData.MDefDownResist);
-        }
-
-        private CDataEquipItemInfo ReadEquipItemInfo(DbDataReader reader)
-        {
-            CDataEquipItemInfo equipItemInfo = new CDataEquipItemInfo();
-            equipItemInfo.ItemId = GetUInt32(reader, "item_id");
-            equipItemInfo.EquipType = GetByte(reader, "equip_type");
-            equipItemInfo.EquipSlot = GetUInt16(reader, "equip_slot");
-            equipItemInfo.Color = GetByte(reader, "color");
-            equipItemInfo.PlusValue = GetByte(reader, "plus_value");
-            return equipItemInfo;
-        }
-
-        private void AddParameter(TCom command, uint characterId, JobId job, CDataEquipItemInfo equipItemInfo)
-        {
-            AddParameter(command, "character_id", characterId);
-            AddParameter(command, "job", (byte) job);
-            AddParameter(command, "item_id", equipItemInfo.ItemId);
-            AddParameter(command, "equip_type", equipItemInfo.EquipType);
-            AddParameter(command, "equip_slot", equipItemInfo.EquipSlot);
-            AddParameter(command, "color", equipItemInfo.Color);
-            AddParameter(command, "plus_value", equipItemInfo.PlusValue);
-        }
-
-        private CDataEquipJobItem ReadEquipJobItem(DbDataReader reader)
-        {
-            CDataEquipJobItem equipJobItem = new CDataEquipJobItem();
-            equipJobItem.JobItemId = GetUInt32(reader, "job_item_id");
-            equipJobItem.EquipSlotNo = GetByte(reader, "equip_slot_no");
-            return equipJobItem;
-        }
-
-        private void AddParameter(TCom command, uint characterId, JobId job, CDataEquipJobItem equipJobItem)
-        {
-            AddParameter(command, "character_id", characterId);
-            AddParameter(command, "job", (byte) job);
-            AddParameter(command, "job_item_id", equipJobItem.JobItemId);
-            AddParameter(command, "equip_slot_no", equipJobItem.EquipSlotNo);
-        }
-
-        private CDataNormalSkillParam ReadNormalSkillParam(DbDataReader reader)
-        {
-            CDataNormalSkillParam normalSkillParam = new CDataNormalSkillParam();
-            normalSkillParam.Job = (JobId) GetByte(reader, "job");
-            normalSkillParam.SkillNo = GetUInt32(reader, "skill_no");
-            normalSkillParam.Index = GetUInt32(reader, "index");
-            normalSkillParam.PreSkillNo = GetUInt32(reader, "pre_skill_no");
-            return normalSkillParam;
-        }
-
-        private void AddParameter(TCom command, uint characterId, CDataNormalSkillParam normalSkillParam)
-        {
-            AddParameter(command, "character_id", characterId);
-            AddParameter(command, "job", (byte) normalSkillParam.Job);
-            AddParameter(command, "skill_no", normalSkillParam.SkillNo);
-            AddParameter(command, "index", normalSkillParam.Index);
-            AddParameter(command, "pre_skill_no", normalSkillParam.PreSkillNo);
-        }
-
-        private CDataSetAcquirementParam ReadSetAcquirementParam(DbDataReader reader)
-        {
-            CDataSetAcquirementParam setAcquirementParam = new CDataSetAcquirementParam();
-            setAcquirementParam.Job = (JobId) GetByte(reader, "job");
-            setAcquirementParam.Type = GetByte(reader, "type");
-            setAcquirementParam.SlotNo = GetByte(reader, "slot_no");
-            setAcquirementParam.AcquirementNo = GetUInt32(reader, "acquirement_no");
-            setAcquirementParam.AcquirementLv = GetByte(reader, "acquirement_lv");
-            return setAcquirementParam;
-        }
-
-        private void AddParameter(TCom command, uint characterId, CDataSetAcquirementParam setAcquirementParam)
-        {
-            AddParameter(command, "character_id", characterId);
-            AddParameter(command, "job", (byte) setAcquirementParam.Job);
-            AddParameter(command, "type", setAcquirementParam.Type);
-            AddParameter(command, "slot_no", setAcquirementParam.SlotNo);
-            AddParameter(command, "acquirement_no", setAcquirementParam.AcquirementNo);
-            AddParameter(command, "acquirement_lv", setAcquirementParam.AcquirementLv);
         }
     }
 }
